fix: 최초 렌더링시 Card의 소유자 이름 및 만료일이 표시되지 않던 버그를 수정한다
